made single tree view with spinner for both subsets and assets
