+++ avalon/tools/views.py
@@ -1,4 +1,6 @@
-from ..vendor.Qt import QtWidgets, QtCore
+import os
+from .. import style
+from ..vendor.Qt import QtWidgets, QtCore, QtGui, QtSvg
 
 
 class DeselectableTreeView(QtWidgets.QTreeView):
@@ -16,7 +18,52 @@ class DeselectableTreeView(QtWidgets.QTreeView):
         QtWidgets.QTreeView.mousePressEvent(self, event)
 
 
-class AssetsView(DeselectableTreeView):
+class TreeViewLoading(QtWidgets.QTreeView):
+    size = 160
+
+    def __init__(self, parent=None):
+        super(TreeViewLoading, self).__init__(parent=parent)
+
+        loading_image_path = os.path.join(
+            os.path.dirname(os.path.abspath(style.__file__)),
+            "svg",
+            "spinner-200.svg"
+        )
+        self.spinner = QtSvg.QSvgRenderer(loading_image_path)
+
+        self.is_loading = False
+        self.is_empty = True
+
+    def paint_loading(self, event):
+        rect = event.rect()
+        rect = QtCore.QRectF(rect.topLeft(), rect.bottomRight())
+        rect.moveTo(
+            rect.x() + rect.width() / 2 - self.size / 2,
+            rect.y() + rect.height() / 2 - self.size / 2
+        )
+        rect.setSize(QtCore.QSizeF(self.size, self.size))
+        painter = QtGui.QPainter(self.viewport())
+        self.spinner.render(painter, rect)
+
+    def paint_empty(self, event):
+        painter = QtGui.QPainter(self.viewport())
+        rect = event.rect()
+        rect = QtCore.QRectF(rect.topLeft(), rect.bottomRight())
+        qtext_opt = QtGui.QTextOption(
+            QtCore.Qt.AlignHCenter | QtCore.Qt.AlignVCenter
+        )
+        painter.drawText(rect, "No Data", qtext_opt)
+
+    def paintEvent(self, event):
+        if self.is_loading:
+            self.paint_loading(event)
+        elif self.is_empty:
+            self.paint_empty(event)
+        else:
+            super(TreeViewLoading, self).paintEvent(event)
+
+
+class AssetsView(TreeViewLoading, DeselectableTreeView):
     """Item view.
     This implements a context menu.
     """
